+++ client/index.js
@@ -1,11 +1,13 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-
+import { HashRouter as Router } from 'react-router-dom'
 import App from './components/App'
 
 document.addEventListener('DOMContentLoaded', () => {
   ReactDOM.render(
-    <App />,
+    <Router>
+      <App />
+    </Router>,
     document.getElementById('app')
   )
 })
