+++ package.json
@@ -60,6 +60,7 @@
     "npm-run-all": "^4.1.5",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-router-dom": "^5.2.0",
     "webpack": "^5.21.0",
     "webpack-cli": "^4.5.0",
     "webpack-dev-server": "^3.11.2"
